requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.2
Werkzeug==3.0.3
fpdf==1.7.2
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
psycopg2-binary==2.9.10
Pillow==10.4.0
